life_assisant_multi_agent/life_assistant/requirements.txt
langchain
chromadb
google-api-python-client
oauth2client
APScheduler
pytest
schedule

